Reverted closings
Reverted back to using a text file on the shared drive. Tampermonkey and Greasemonkey cache @resource and @required. Changes to the version hosted on GitHub would not update frequently enough to work. You can change the update interval in Tampermonkey, but this would also grab the jQuery lib at the same frequency.